+++ Assets/Scripts/Gameplay/Services/ConstructionElements/CapitalService.cs
@@ -6,10 +6,10 @@ namespace Gameplay.Services.ConstructionElements
 {
     public class CapitalService : ICapitalService
     {
-        private float _unitsCapital;
+        private int _unitsCapital;
         private int _gold;
 
-        private float _unitsPerSecond = 1f;
+        private int _unitsPerSecond = 1;
         
         private int _goldPerSecond = 1;
 
@@ -25,6 +25,6 @@ namespace Gameplay.Services.ConstructionElements
             AddUnits?.Invoke(_unitsCapital);
         }
         
-        public float GetUnits() => _unitsCapital;
+        public int GetUnits() => _unitsCapital;
     }
 }
